Add inline weapon command
- Refactor the effects classes for clarity
- Fix effects that are not statuses not being cleared by Clear
