+++ lang/en.json
@@ -131,6 +131,7 @@
 		"InlineDamage": "Apply damage to selected token",
 		"InlineRecovery": "Apply resource recovery to selected token",
 		"InlineLoss": "Apply resource loss to selected tokens",
+		"InlineWeapon": "Apply effect to equipped weapon on character",
 		"Name": "Name",
 		"Roll": "Roll",
 		"Add": "Add",
@@ -839,6 +840,7 @@
 		"IgnoreNone": "Ignore None",
 		"ChatApplyTargeted": "Apply Targeted",
 		"ChatApplySelected": "Apply Selected",
+		"ChatApplyNoCharacterSelected": "No character selected. Unable to apply.",
 		"ChatApplySelectedTooltip": "Click to apply effect to selected tokens.",
 		"ChatApplyEffectNoActorsSelected": "No actors selected. Unable to apply effect.",
 		"ChatApplyEffectNoActorsTargeted": "No actors targeted. Unable to apply effect.",
